[refactor]:Socket 파일 이동
